requirements.txt
appium-python-client>=3.1.0
selenium>=4.18.0
pytest>=8.0.0
hypothesis>=6.100.0
allure-pytest>=2.13.0
python-dotenv>=1.0.0
